<[Epic-ID: ODUHIGH-406][Task-ID: ODUHIGH-421]Paging Message: CU_STUB Trigger and DUAPP Handling>

+++ src/codec_utils/common/odu_common_codec.h
@@ -28,7 +28,7 @@ char encBuf[ENC_BUF_MAX_LEN];
 int  encBufSize;
 
 int PrepFinalEncBuf(const void *buffer, size_t size, void *encodedBuf);
-uint8_t fillBitString(BIT_STRING_t *id, uint8_t unusedBits, uint8_t byteSize, uint8_t val);
+uint8_t fillBitString(BIT_STRING_t *id, uint8_t unusedBits, uint8_t byteSize, uint64_t val);
 uint8_t bitStringToInt(BIT_STRING_t *bitString, void *val);
 void fillTeIdString(uint8_t bufSize, uint32_t val, uint8_t *buf);
 void teIdStringToInt(uint8_t *buf, uint32_t *val);
